refactor(core): share the interactable persistence scheduler
