+++ .changeset/interactables-persistence-queue.md
@@ -1,0 +1,5 @@
+---
+"@assistant-ui/core": patch
+---
+
+refactor: share the interactable persistence scheduler between the tap client and the legacy surface.
